feat(approach): PiP leads with the radar station price + km while still approaching
Surfaces the Fuel Station Radar in the PiP small window EARLIER than the
geo-fence flip: while still approaching (ApproachPolling) the tile now leads
with the nearest priced radar station's price + km instead of consumption.

- TripRecordingPipView gains radarStation / radarDistanceMeters / radiusMeters
  (paint-only). Layout selector: in-radius/leaving → price (metres) ; else
  radarStation != null → price (km) ; else → consumption (never blank).
- The two price layouts are unified in the extracted
  TripRecordingPipPriceLayout (km vs metres caption flag) + the proximity bar,
  keeping the #2620 single-FittedBox no-clip host and #2601 tap-to-navigate.
- TripRecordingBanner resolves nearestStationRadarProvider + the radar radius
  (each guarded) and passes them in; the fuel-hue palette now also covers the
  polling-with-radar case. approachOverlayPalette → fuelStationRadarPalette.
- The banner content row is extracted to TripRecordingBannerContent to keep
  the banner file under the 400-line cap after the new plumbing.

PiP tests cover: radar price + km in polling, fallback to consumption when
null, in-radius still wins, and the green proximity bar.

Closes #2661.
